requirements.txt
beautifulsoup4==4.3.2
githubpy==1.1.0
httplib2==0.9
tabulate==0.7.3
